Refactor settings components
